requirements.txt
aiohttp
python-dotenv
uvloop
rapidjson